Finished styling bookings and calendar
Formatted and displayed bookings, styled appropriately. Bookings are sorted out by their start dates in the Bookings page component, then use several BookingCard components to format and display the information for the specified booking.

Implemented dynamic styling for the calendar tiles based on upcoming bookings. Declined bookings do not show on the calendar. This specific styling proved to be more elegant than the previous way to highlight, so adjusted the active styling elsewhere as well.

Adjusted props in several components as needed.