API Data for Bookings
Returning data from the API for bookings. The GET method for requests now also retrieves user profiles for every request.

Changed startTime/endTime to startDate/endDate in the request interface and relevant areas.

Changed the date format to be more descriptive and readable.

Moved many of the values to useState() to trigger update on API load.

Fixed typo in the user controller.