Created RequestStatusButton
Encapsulated the SettingsIcon in a RequestStatusButton component. This component will display a dropdown menu when clicked that changes the request status.

Added the changeRequestStatus API helper that takes in an id and status.

Removed the userType param and query from getRequests helper.

Added an id to the Request interface for editing functionality.

Fixed calendar active class on dates with single digits.

Fixed typo in Request model