Cleaning up and Snackbar
Added messages to the Snackbar on request status change.

Used clsx to clean up some code. Refactored the formatted date to use string literals.

Streamlined the RequestStatusButton. Moved onStatusChange directly to the handlers instead of relying on state change to call it in useEffect(), especially since the state was unnecessary in the first place.

Used .push instead of assigning a value to a specific element in an array by index when populating the userBookings object in Bookings.