UI updates
Used variants in Typography and standardized their font-sizes through the theme. This is subject to change.

Added error color to the theme.

Adjusted max size of the grid on different screen sizes to maintain aspect ratio. Added window width to the state and an event listener to set it so the page is more responsive on resize.

Added the parameter version of the route to App.tsx.

Fixed an explicit any type.