Reviews FE
Prepared ProfileDetails front end for reviews. There is a preview section for up to three reviews and, if there are any more, a button that brings up a modal to show all reviews.

Added a Review component, ReviewDialog (modal) component, and Review interface.

Made a small adjustment to review controller to prepare for populating name and photo url.