Password Reset integration
Set up password reset form at /password-reset/:email/:id

Added ResetToken model to the database.

The back end receives the token and email that are in the URL params on navigating to the page along with the new password. The token is checked against a hashed token in the database. Tokens created by sending a password reset email only last 30 minutes. After password is changed, all tokens belonging to the user are deleted.

The send password reset function now sends the email in the link for convenience (instead of having the user enter it again)